+++ app/components/product-form/placeholder.tsx
